services/api/app/requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
cassandra-driver
